# static/swagger.yaml
openapi: 3.0.0
info:
  title: Cloud Resource Autoscaler API
  description: |
    API for monitoring and auto-scaling cloud resources based on metrics.
    
    ## Features
    - Mock instances for testing without AWS
    - Time-based scaling with 5-minute sustained usage checks
    - Instance capacity tracking
    - Prolonged metric simulation
    
    ## Authentication
    Most endpoints require a JWT token obtained from the login endpoint.
    Use the "Authorize" button to set your Bearer token.
  version: 2.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        instance_count:
          type: integer
        monitoring_count:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: string
                    format: uuid
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Authenticate and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get user information
      description: Get current authenticated user's profile and statistics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...